added StatsdPerformanceMeasure for publishing metrics from within a using block
